+++ app/build.gradle
@@ -50,6 +50,7 @@ dependencies {
     implementation 'com.firebaseui:firebase-ui-database:4.3.2'
     implementation 'com.google.firebase:firebase-core:16.0.8'
     implementation 'com.google.firebase:firebase-firestore:18.2.0'
+
     // Google Sign In SDK (only required for Google Sign In)
     //implementation 'com.google.android.gms:play-services-auth:16.0.1'
     implementation 'android.arch.lifecycle:extensions:1.1.1'
