+++ app/src/main/AndroidManifest.xml
@@ -22,7 +22,7 @@
         <activity android:name=".activities.Graphs.MyPieChartActivity" />
         <activity android:name=".activities.Graphs.MyBarGraph" />
         <activity android:name=".activities.Graphs.MyLineGraph" />
-        <activity android:name=".Notifications.NotificationActivity">
+        <activity android:name=".notifications.NotificationActivity">
             <intent-filter>
                 <action android:name="com.sda5.walletdroid.TargetNotification" />
 
@@ -33,8 +33,6 @@
 
         <activity android:name=".activities.ExpenseDetailsActivity"></activity>
 
-
-        <activity android:name=".activities.ExpenseDetailsActivity"></activity>
 
         <activity android:name=".activities.SeeExpenseGraphForParticularCategory"></activity>
         <activity android:name=".activities.QueryActivity" />
@@ -82,11 +80,15 @@
             android:name="com.google.android.gms.ads.AD_MANAGER_APP"
             android:value="true" />
 
-        <service android:name=".Notifications.FirebaseMessagingService">
+        <service android:name=".notifications.FirebaseMessagingService">
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT" />
+
             </intent-filter>
         </service>
+        <meta-data
+            android:name="com.google.firebase.messaging.default_notification_channel_id"
+            android:value="@string/default_notification_channel_id"/>
     </application>
 
 </manifest>
