+++ app/src/main/java/com/sda5/walletdroid/activities/ServiceActivity.java
@@ -18,13 +18,10 @@ import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.google.android.material.navigation.NavigationView;
 import com.google.android.material.textfield.TextInputLayout;
 import com.google.firebase.auth.FirebaseAuth;
-import com.sda5.walletdroid.Notifications.NotificationFragment;
+import com.sda5.walletdroid.notifications.NotificationFragment;
 import com.sda5.walletdroid.R;
 import com.sda5.walletdroid.fragments.ExpenseFragment;
 import com.sda5.walletdroid.fragments.GroupFragment;
-import com.sda5.walletdroid.fragments.InvestFragment;
-import com.sda5.walletdroid.fragments.QueyFragment;
-import com.sda5.walletdroid.fragments.SettleFragment;
 import com.sda5.walletdroid.fragments.fragments_navigation.AppDetails;
 import com.sda5.walletdroid.fragments.fragments_navigation.Feedbacknav;
 import com.sda5.walletdroid.fragments.fragments_navigation.GraphFragment;
