+++ app/src/main/java/com/sda5/walletdroid/notifications/FirebaseMessagingService.java
@@ -1,0 +1,81 @@
+package com.sda5.walletdroid.notifications;
+
+import android.app.NotificationChannel;
+import android.app.NotificationManager;
+import android.app.PendingIntent;
+import android.content.Intent;
+import android.graphics.Color;
+import android.media.RingtoneManager;
+import android.net.Uri;
+
+import androidx.core.app.NotificationCompat;
+
+import com.google.firebase.messaging.RemoteMessage;
+import com.sda5.walletdroid.R;
+
+public class FirebaseMessagingService extends com.google.firebase.messaging.FirebaseMessagingService {
+
+    private static final String TAG = "FirebaseService";
+    public static final String NOTIFICATION_CHANNEL_ID = "11111";
+
+    @Override
+    public void onMessageReceived(RemoteMessage remoteMessage) {
+        super.onMessageReceived(remoteMessage);
+
+        String click_Action = remoteMessage.getNotification().getClickAction();
+        String message = remoteMessage.getData().get("message");
+
+        String dataFrom = remoteMessage.getData().get("sentBy");
+
+        String messageTitle="Walletdroid Settlement/Expence Notice";
+
+        //String messageTitle = remoteMessage.getNotification().getTitle();
+        String messageBody = remoteMessage.getNotification().getBody();
+
+        Uri defaultSoundUri = RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION);
+
+
+        NotificationCompat.Builder mBuilder = new NotificationCompat.Builder(this, getString(R.string.default_notification_channel_id))
+                .setSmallIcon(R.mipmap.ic_launcher)
+                .setContentTitle(messageTitle)
+                .setContentText(messageBody)
+                .setSound(defaultSoundUri)
+                .setPriority(NotificationCompat.PRIORITY_HIGH);
+        int mNotificationID = (int) System.currentTimeMillis();
+        NotificationManager manager = (NotificationManager) getSystemService(NOTIFICATION_SERVICE);
+
+        if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.O)
+        {
+            int importance = NotificationManager.IMPORTANCE_HIGH;
+            NotificationChannel notificationChannel = new NotificationChannel(NOTIFICATION_CHANNEL_ID, "NOTIFICATION_CHANNEL_NAME", importance);
+            notificationChannel.enableLights(true);
+            notificationChannel.setLightColor(Color.RED);
+            notificationChannel.enableVibration(true);
+            notificationChannel.setVibrationPattern(new long[]{100, 200, 300, 400, 500, 400, 300, 200, 400});
+            assert manager != null;
+            mBuilder.setChannelId(NOTIFICATION_CHANNEL_ID);
+            manager.createNotificationChannel(notificationChannel);
+        }
+
+
+
+        Intent intent = new Intent(click_Action);
+        intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
+        intent.putExtra("message", message);
+        //intent.putExtra("from_user_id",dataFrom);
+        intent.putExtra("sentBy", dataFrom);
+        PendingIntent pendingIntent = PendingIntent.getActivity(this, 0, intent, PendingIntent.FLAG_UPDATE_CURRENT);
+        mBuilder.setContentIntent(pendingIntent);
+
+        manager.notify(mNotificationID, mBuilder.build());
+
+
+
+    }
+
+
+
+
+
+}
+
